+++ src/frontend/apps/impress/src/features/docs/doc-header/components/DocHeaderInfo.tsx
@@ -1,3 +1,4 @@
+import { Tooltip } from '@gouvfr-lasuite/ui-components';
 import { t } from 'i18next';
 
 import { Box, Icon, Text } from '@/components';
@@ -22,10 +23,11 @@ interface DocHeaderInfoProps {
 export const DocHeaderInfo = ({ doc }: DocHeaderInfoProps) => {
   const { transRole } = useTrans();
   const { isEditable } = useIsCollaborativeEditable(doc);
-  const { relativeDate, calculateDaysLeft } = useDate();
+  const { relativeDate, formatDate, calculateDaysLeft } = useDate();
   const { data: config } = useConfig();
 
   const relativeOnly = relativeDate(doc.updated_at);
+  const fullDate = formatDate(doc.updated_at);
 
   const trashbinCutoff = config?.TRASHBIN_CUTOFF_DAYS;
 
@@ -87,8 +89,21 @@ export const DocHeaderInfo = ({ doc }: DocHeaderInfoProps) => {
         {dateLabel}
         &nbsp;
       </Text>
-      <Text as="dd" $variation="tertiary" $size="s" $margin="0">
-        {dateValue}
+      <Text
+        as="dd"
+        $variation="tertiary"
+        $size="s"
+        $direction="row"
+        $align="center"
+        $margin="0"
+      >
+        {trashbinCutoff && doc.deleted_at ? (
+          dateValue
+        ) : (
+          <Tooltip content={fullDate} placement="top">
+            <time dateTime={doc.updated_at}>{relativeOnly}</time>
+          </Tooltip>
+        )}
       </Text>
     </Box>
   );